backend/requirements.txt
fastapi==0.112.2
uvicorn[standard]==0.30.6
pydantic==2.9.2
python-dotenv==1.0.1
httpx==0.27.2
pydantic-settings==2.5.2
opencv-python==4.10.0.84
mediapipe==0.10.14

